+++ xrpl/queries/account/transactions_test.go
@@ -30,3 +30,58 @@ func TestAccountTransactionsRequest(t *testing.T) {
 		t.Error(err)
 	}
 }
+
+func TestAccountTransactionsResponse(t *testing.T) {
+	s := TransactionsResponse{
+		Account: "abc",
+		LedgerIndexMin: 100,
+		LedgerIndexMax: 120,
+		Limit: 10,
+		Marker: "123",
+		Transactions: []Transaction{
+			{
+				Hash: "def",
+				LedgerHash: "ghi",
+				LedgerIndex: 10,
+				CloseTimeISO: "2021-01-01T00:00:00Z",
+				Tx: map[string]any{
+					"TransactionType": "Payment",
+					"Account": "abc",
+					"Destination": "def",
+					"Amount": "100",
+				},
+			},
+		},
+	}
+
+	j := `{
+	"account": "abc",
+	"ledger_index_min": 100,
+	"ledger_index_max": 120,
+	"limit": 10,
+	"marker": "123",
+	"transactions": [
+		{
+			"close_time_iso": "2021-01-01T00:00:00Z",
+			"hash": "def",
+			"ledger_hash": "ghi",
+			"ledger_index": 10,
+			"meta": {},
+			"tx_json": {
+				"Account": "abc",
+				"Amount": "100",
+				"Destination": "def",
+				"TransactionType": "Payment"
+			},
+			"tx_blob": "",
+			"validated": false
+		}
+	],
+	"validated": false
+}`
+
+	if err := testutil.Serialize(t, s, j); err != nil {
+		t.Error(err)
+	}
+}
+